rfr: move Blog and Comment to packages
