+++ api/src/main/java/ch/heig/dai/lab/http/api/blog/Blog.java
@@ -1,4 +1,4 @@
-package ch.heig.dai.lab.http.api;
+package ch.heig.dai.lab.http.api.blog;
 
 /**
  * Blog model.
