+++ api/src/main/java/ch/heig/dai/lab/http/api/blog/BlogController.java
@@ -1,4 +1,4 @@
-package ch.heig.dai.lab.http.api;
+package ch.heig.dai.lab.http.api.blog;
 
 import io.javalin.apibuilder.CrudHandler;
 import io.javalin.http.Context;
